Django_api/requirements.txt
Django
djangorestframework
gunicorn
mysqlclient
djangorestframework-simplejwt
python-dotenv
django-cors-headers
drf-yasg
pymongo
celery
amqp
pytest
requests